oops money fix

+++ Blank TD/Assets/Scripts/MoneyManager.cs
@@ -11,7 +11,7 @@ public class MoneyManager : MonoBehaviour
     public Text moneyDisplay;
     public Text sellButtonDisplay;
     public Text buyPriceDisplay;
-    public int money;
+    public float money;
     public float sellPrice;
     public int buyPrice;
     public bool noFakeStart;
